fix world state problem
