+++ ethereum/blockcreation/src/test/java/org/hyperledger/besu/ethereum/blockcreation/AbstractBlockTransactionSelectorTest.java
@@ -45,7 +45,6 @@ import org.hyperledger.besu.ethereum.eth.transactions.ImmutableTransactionPoolCo
 import org.hyperledger.besu.ethereum.eth.transactions.PendingTransactions;
 import org.hyperledger.besu.ethereum.eth.transactions.sorter.BaseFeePendingTransactionsSorter;
 import org.hyperledger.besu.ethereum.mainnet.MainnetTransactionProcessor;
-import org.hyperledger.besu.ethereum.mainnet.MainnetTransactionValidator;
 import org.hyperledger.besu.ethereum.mainnet.ProtocolSchedule;
 import org.hyperledger.besu.ethereum.mainnet.ValidationResult;
 import org.hyperledger.besu.ethereum.mainnet.feemarket.FeeMarket;
@@ -57,10 +56,8 @@ import org.hyperledger.besu.evm.gascalculator.LondonGasCalculator;
 import org.hyperledger.besu.evm.internal.EvmConfiguration;
 import org.hyperledger.besu.evm.worldstate.WorldState;
 import org.hyperledger.besu.metrics.noop.NoOpMetricsSystem;
-import org.hyperledger.besu.plugin.data.TransactionSelectionResult;
 import org.hyperledger.besu.plugin.data.TransactionType;
 import org.hyperledger.besu.plugin.services.MetricsSystem;
-import org.hyperledger.besu.plugin.services.txselection.TransactionSelector;
 import org.hyperledger.besu.testutil.TestClock;
 
 import java.math.BigInteger;
@@ -85,22 +82,12 @@ public abstract class AbstractBlockTransactionSelectorTest {
   protected static final double MIN_OCCUPANCY_100_PERCENT = 1;
   protected static final KeyPair keyPair =
       SignatureAlgorithmFactory.getInstance().generateKeyPair();
-  public static final TransactionSelector DO_NOTHING_TRANSACTION_SELECTOR =
-      new TransactionSelector() {
-        @Override
-        public TransactionSelectionResult selectTransaction(
-            final org.hyperledger.besu.plugin.data.Transaction transaction,
-            final org.hyperledger.besu.plugin.data.TransactionReceipt receipt) {
-          return TransactionSelectionResult.CONTINUE;
-        }
-      };
   protected final MetricsSystem metricsSystem = new NoOpMetricsSystem();
 
   protected final Blockchain blockchain = new ReferenceTestBlockchain();
   protected PendingTransactions pendingTransactions;
   protected MutableWorldState worldState;
   @Mock protected MainnetTransactionProcessor transactionProcessor;
-  @Mock protected MainnetTransactionValidator transactionValidator;
 
   @Before
   public void setup() {
